refactor : 빼빼로 구현
